requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
scikit-learn
nltk 